下一張圖
修正Gallery的檔案列表

+++ EABook/drawing/EADraw.h
@@ -21,7 +21,7 @@
     NSArray *cavas;
     
     BOOL drawAble;
-    NSInteger canvasImageNum;
+    int canvasImageNum;
     NSUInteger SelectedCrayon;
     
     DrawGallery *gallery;
